more apis added, misc apis file added, need testing
