+++ src/components/pages/referralCode.js
@@ -129,7 +129,7 @@ class ReferralCode extends Component {
 
         // make the loading bar show up
         self.setState({...this.state, loading: true});
-        axios.post("api/createReferralCode", {name, email})
+        axios.post("api/misc/createReferralCode", {name, email})
         .then(function(response) {
             // stop the loading bar, show the referral code
             const referralCode = response.data;
